// cross-chain-sdk/src/api/orders/types.ts
export enum FillStatus {
    Pending = 'pending',
    Executed = 'executed',
    Refunding = 'refunding',
    Refunded = 'refunded'
}

export enum OrderStatus {
    Pending = 'pending',
    Executed = 'executed',
    Expired = 'expired',
    Cancelled = 'cancelled',
    Refunding = 'refunding',
    Refunded = 'refunded'
}

export type Fill = {
    status: FillStatus
    txHash: string
    filledMakerAmount: string
    filledAuctionTakerAmount: string
    escrowEvents: EscrowEventData[]
}

export enum EscrowEventSide {
    Src = 'src',
    Dst = 'dst'
}

export enum EscrowEventAction {
    SrcEscrowCreated = 'src_escrow_created',
    DstEscrowCreated = 'dst_escrow_created',
    Withdrawn = 'withdrawn',
    FundsRescued = 'funds_rescued',
    EscrowCancelled = 'escrow_cancelled'
}

export type EscrowEventData = {
    transactionHash: string
    escrow: string
    side: EscrowEventSide
    action: EscrowEventAction
    blockTimestamp: number
}

export type OrderStatusResponse = {
    status: OrderStatus
    order: LimitOrderV4Struct
    extension: string
    points: AuctionPoint[] | null
    cancelTx: string | null
    fills: Fill[]
    createdAt: string
    auctionStartDate: number
    auctionDuration: number
    initialRateBump: number
    isNativeCurrency: boolean
    fromTokenToUsdPrice: string
    toTokenToUsdPrice: string
}
// ...
export type ReadyToAcceptSecretFill = {
    idx: number
    srcEscrowDeployTxHash: string
    dstEscrowDeployTxHash: string
}

export type ReadyToAcceptSecretFills = {
    fills: ReadyToAcceptSecretFill[]
}